+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PR [#147](https://github.com/marinasundstrom/CheckedExceptions/pull/147) Disallow prop decl with Throws and Get accessor
 
+### Fixed
+
+- PR [#148](https://github.com/marinasundstrom/CheckedExceptions/pull/148) Handle inheritance with props w Throws and get
+
 ## [1.6.3] - 2025-07-28
 
 ### Added